refactor: hard-coded keys
