+++ frontend/src/components/auth/SignUp/SignUpOptionsForm.tsx
@@ -6,7 +6,7 @@ import Button from '@/components/Primitives/Inputs/Button/Button';
 import Flex from '@/components/Primitives/Layout/Flex/Flex';
 import Text from '@/components/Primitives/Text/Text';
 import { SignUpEnum } from '@/utils/signUp.enum';
-import loginWithAzure from '@/hooks/users/loginWithAzure';
+import loginWithAzure from '@/hooks/auth/loginWithAzure';
 import { OrSeparator } from '../LoginForm/styles';
 
 const Container = styled(Flex, { width: '100%' });
