+++ frontend/src/hooks/auth/useRegisterGuestUser.tsx
