+++ frontend/src/hooks/teams/useTeamsWithoutUser.tsx
@@ -7,11 +7,12 @@ import { useSetRecoilState } from 'recoil';
 import { toastState } from '@/store/toast/atom/toast.atom';
 
 import { TEAMS_KEY } from '.';
+import { USERS_KEY } from '../users';
 
 const useTeamsWithoutUser = (userId: string) => {
   const setToastState = useSetRecoilState(toastState);
 
-  return useQuery([TEAMS_KEY, 'not', 'user', userId], () => getTeamsWithoutUser(userId), {
+  return useQuery([TEAMS_KEY, 'not', USERS_KEY, userId], () => getTeamsWithoutUser(userId), {
     enabled: true,
     refetchOnMount: false,
     refetchOnWindowFocus: false,
